+++ changes.md
@@ -1,7 +1,11 @@
 Monomer Changes
 ===============
 
+** November 9, 2013
+
+* Implemented Button, PostButton and DeleteButton.
+
 ** November 6, 2013
 
 * First project draft.
-* Added Button component.
+* Added Component.